requirement.txt
pinecone
openai
python-dotenv
haystack-ai
requests


pytelegrambotapi
docling
loguru